Add back info circles, remove underlines
