+++ ts/deckoptions/StepsInputRow.svelte
@@ -3,11 +3,9 @@
     License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="ts">
-    import marked from "marked";
     import Row from "./Row.svelte";
     import Col from "./Col.svelte";
-    import WithTooltip from "./WithTooltip.svelte";
-    import Label from "./Label.svelte";
+    import TooltipLabel from "./TooltipLabel.svelte";
     import StepsInput from "./StepsInput.svelte";
     import RevertButton from "./RevertButton.svelte";
 
@@ -18,11 +16,7 @@
 
 <Row>
     <Col size={7}>
-        <WithTooltip tooltip={marked(markdownTooltip)} let:createTooltip>
-            <Label on:mount={(event) => createTooltip(event.detail.span)}
-                ><slot /></Label
-            >
-        </WithTooltip>
+        <TooltipLabel {markdownTooltip}><slot /></TooltipLabel>
     </Col>
     <Col size={5}>
         <RevertButton bind:value {defaultValue} />
